+++ src/lrcalc/func.rs
@@ -1,12 +1,10 @@
-use super::helper::{
-    part::PartitionIterator,
-    perm::{all_perms as _all_perms, all_strings as _all_strings},
-};
+use super::helper::part::PartitionIterator;
 
 use super::{
     ivector::{ivl_to_vec, IntVector},
     lincomb::LinearCombination,
     lriter::LRTableauIterator,
+    perm::{all_perms as _all_perms, all_strings as _all_strings},
     schublib::{mult_poly_schubert, mult_schubert, mult_schubert_str, trans},
     schur::{
         fusion_reduce_lc, schur_coprod, schur_lrcoef, schur_mult, schur_mult_fusion, schur_skew,
