+++ src/lrcalc/helper/mod.rs
@@ -1,4 +1,3 @@
 pub mod ivector;
 pub mod ivlincomb;
 pub mod part;
-pub mod perm;
