+++ src/lrcalc/schublib.rs
@@ -4,9 +4,10 @@ use super::helper::{
         _ivlc_insert, ivlc_add_element, ivlc_add_multiple, ivlc_new_default, LinearCombination,
         LC_FREE_ZERO,
     },
-    perm::{
-        bruhat_zero, perm2string, perm_group, perm_length, str2dimvec, str_iscompat, string2perm,
-    },
+};
+
+use super::perm::{
+    bruhat_zero, perm2string, perm_group, perm_length, str2dimvec, str_iscompat, string2perm,
 };
 
 pub fn trans(w: &[i32], vars: i32) -> LinearCombination {
