+++ src/schubmult.rs
@@ -4,8 +4,7 @@ use anyhow::{ensure, Context, Result};
 use clap::AppSettings::{AllowLeadingHyphen, DeriveDisplayOrder};
 use clap::{crate_version, Clap};
 
-use lrcalc::helper::perm::str_iscompat;
-use lrcalc::{is_permutation, schubmult, schubmult_str};
+use lrcalc::{is_permutation, perm::str_iscompat, schubmult, schubmult_str};
 
 fn check_non_negative(num_str: &str) -> Result<()> {
     let num = num_str
